pyproject.toml
[project]
name = "workmy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.6.2",
    "django-ninja-jwt>=5.4.4",
    "gunicorn>=25.3.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "whitenoise>=6.12.0",
]
